+++ Editor/SerializedType.cs
@@ -4,16 +4,23 @@ using UnityEngine;
 namespace Unity.Tutorials.Core.Editor
 {
     /// <summary>
-    /// Used to serialize System.Type.
+    /// Used to serialize System.Type using Type.AssemblyQualifiedName.
     /// </summary>
     [Serializable]
     public class SerializedType : ISerializationCallbackReceiver
     {
-        [SerializeField]
-        string m_TypeName;
+        static readonly bool k_IsAuthoringMode = ProjectMode.IsAuthoringMode();
 
         /// <summary>
-        /// The Type.
+        /// Is the type specified, meaning, the assembly-qualified type name is stored.
+        /// </summary>
+        internal bool IsSpecified => m_TypeName.IsNotNullOrEmpty();
+
+        [SerializeField]
+        internal string m_TypeName;
+
+        /// <summary>
+        /// The Type is stored using its assembly-qualified type name.
         /// </summary>
         public Type Type
         {
@@ -27,7 +34,7 @@ namespace Unity.Tutorials.Core.Editor
         /// <param name="type"></param>
         public SerializedType(Type type)
         {
-            this.Type = type;
+            Type = type;
         }
 
         /// <summary>
@@ -44,7 +51,7 @@ namespace Unity.Tutorials.Core.Editor
         {
             // TODO what's this? Is this needed?
             // Remove "-testable" suffix from assembly names
-            if (!ProjectMode.IsAuthoringMode() && m_TypeName != null)
+            if (!k_IsAuthoringMode && IsSpecified)
             {
                 m_TypeName = m_TypeName.Replace("Assembly-CSharp-Editor-firstpass-testable", "Assembly-CSharp-Editor-firstpass");
                 m_TypeName = m_TypeName.Replace("Assembly-CSharp-Editor-testable", "Assembly-CSharp-Editor");
@@ -54,11 +61,22 @@ namespace Unity.Tutorials.Core.Editor
 
             // Backwards-compatibility for IET < 2.0 when the namespace and assemblies were Unity.InteractiveTutorials.*
             // instead of Unity.Tutorials.Core(.Editor).
-            if (m_TypeName.IsNotNullOrEmpty())
+            if (IsSpecified)
             {
                 m_TypeName = m_TypeName.Replace("Unity.InteractiveTutorials.Core", "Unity.Tutorials.Core.Editor");
                 m_TypeName = m_TypeName.Replace("Unity.InteractiveTutorials", "Unity.Tutorials.Core.Editor");
             }
+
+            // TODO figure out how to log these issues: we don't want to log warnings/errors always, and especially
+            // without showing which is the asset causing the warnings/errors, as sometimes the author might want to keep
+            // missing types when authoring assets for a project which supports multiple Unity versions.
+            //if (ProjectMode.IsAuthoringMode() && IsSpecified)
+            //{
+            //    if (Type == null)
+            //        Debug.LogError($"Type name '{m_TypeName}' was specified but it could not been resolved into a run-time Type.");
+            //    else if (Type.AssemblyQualifiedName != m_TypeName)
+            //        Debug.LogWarning($"Type name '{m_TypeName}' was resolved into a run-time Type '{Type}' but with a different type name '{Type.AssemblyQualifiedName}'");
+            //}
         }
     }
 }