+++ Documentation~/whats-new.md
@@ -6,6 +6,11 @@ Summary of changes in Tutorial Framework package version 2.1.
 
 The main updates in this release include:
 
+### Changed
+- When traversing back to an already completed tutorial page, the masking settings of the page are reapplied if the page doesn't have completion criteria set.
+This new behavior improves the tutorial authoring experience and allows a tutorial user to reobserve pages in "a tour of the UI" type of tutorials.
+- When applying masking, do not not throw an exception if unmasked view's **Editor Window Type** or **View Type** is explicitly set to **None**, meaning, the masking is likely not fully configured yet.
+
 ### Added
 - Added support for UI Toolkit masking and highlighting.
 - Added multiple scene support for tutorials: the first element of **Scenes** list is considered to be the main scene and the rest of scenes are loaded additively.
@@ -15,7 +20,9 @@ The main updates in this release include:
 - Scripting API: Made `TutorialModalWindow` part of the public API of the package. This class can be used to implement welcome/closing dialogs for the tutorial project.
 
 ### Fixed
-- Fixed misplaced unmasking of UI controls on Unity 2021.2.0.
+- Fixed unresolved `SerializedType` values being shown incorrectly as **None**. These values are now shown as **Not Found** with a red background in the Inspector.
+- Fixed `SerializedType` values that were resolved but the resolved type had a different assembly-qualified type name than the original name being shown incorrectly as **None**. These values are now shown with a yellow background in the Inspector.
+- Fixed misplaced unmasking of UI controls in floating editor windows on Unity 2021.2.0.
 - Fixed memory leak errors ("A Native Collection has not been disposed...") that occurred on Unity 2021.
 - UI: Disabled horizontal scrollbars in all windows.
 - Fixed original scenes not being restored correctly when exiting a tutorial which contained multiple scenes.